refactor: 함수 및 클래스 분리
