+++ src/main/kotlin/racingcar/View/InputView.kt
@@ -1,0 +1,11 @@
+package racingcar.View
+
+class InputView {
+    fun inputCars() {
+        println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)")
+    }
+
+    fun inputCount() {
+        println("시도할 횟수는 몇 회인가요?")
+    }
+}
